Store the log and .done file into /file
The volumes /file and /config don't have the same lifecycle.

- /config, being defined in the image is recreated at each container creation
- /file, being defined outside of the image, is kept between containir creations

This means that recreating the container, as it is common thing to
do in k8s, will recreated the /config volume while keeping the
/file volume.

Because the curren logic was to put the vault initialized state in
/config, a restart of vault will make the init script think vault
is not initialized, run the initialization and crash because vault
realizes there are stuffs in /file.

By putting the state as long as the files in /file, the vault
initialization state is ensured to be coherent with the files in
/file.